Fixed overwrite bug
The script used to save all weeks of a month under the same name, which overwrote each new week that was being fetched. The name only changed each month, meaning that only the last week of a month was being fetched, instead of all the weeks of a month.

I know, it's confusing